requirements.txt
absl-py==1.2.0
aio-pika==8.2.4
aiofiles==22.1.0
aiogram==2.19
aiohttp==3.8.3
aiormq==6.4.2
aiosignal==1.2.0
aniso8601==9.0.1
APScheduler==3.9.1
argon2-cffi==21.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bidict==0.22.0
bleach==4.1.0
blis==0.7.9
boto3==1.26.2
botocore==1.29.2
CacheControl==0.12.11
cachetools==5.2.0
catalogue==2.0.8
certifi #@ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.0
charset-normalizer==2.1.1
cli-helpers==2.1.0
click==8.0.1
cloudpickle==2.2.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.1.0
confection==0.0.3
configobj==5.0.6
cryptography==38.0.3
cycler==0.10.0
cymem==2.0.7
dask==2022.2.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.0.12
fire==0.4.0
Flask==2.0.1
Flask-RESTful==0.3.9
flatbuffers==22.10.26
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gast==0.5.3
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==2.0.0.post0
grpcio==1.50.0
h5py==3.7.0
httplib2==0.21.0
httptools==0.5.0
huggingface-hub==0.10.1
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.2.0
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kafka-python==2.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==14.0.6
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mattermostwrapper==2.2
mistune==0.8.4
msgpack==1.0.4
multidict==5.2.0
murmurhash==1.0.9
mysql-connector==2.2.9
mysql-connector-python==8.0.31
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.21.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
partd==1.3.0
pathy==0.6.2
pendulum==2.1.2
pgcli==3.1.0
pgspecial==1.13.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.2.0
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.19.6
psycopg2==2.9.1
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.10.2
pydot==1.4.2
Pygments==2.9.0
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
questionary==1.10.0
randomname==0.1.5
rasa==3.3.0
rasa-sdk==3.3.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.26.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacremoses==0.0.53
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.9.10
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==5.2.1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.42
sqlparse==0.4.1
srsly==2.4.5
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.3
tensorflow-addons==0.17.1
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-text==2.8.2
termcolor==2.1.0
terminado==0.12.1
terminaltables==3.1.0
testpath==0.5.0
thinc==8.1.5
threadpoolctl==3.0.0
tokenizers==0.10.3
toolz==0.12.0
torch==1.10.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.0
transformers==4.13.0
twilio==7.14.2
typeguard==2.13.3
typer==0.4.2
typing-utils==0.1.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
uvloop==0.17.0
virtualenv==20.6.0
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
